pyproject.toml
[tool.poetry]
name = "gamepop-bot"
version = "0.1.0"
description = "The Theme Voting Bot for The Alberta GamePop Game Jam"
authors = ["Matthew Hrehirchuk <mhreh594@mtroyal.ca>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.group.dev.dependencies]
ruff = "^0.4.8"
alembic = {extras = ["tz"], version = "^1.13.2"}

[tool.ruff.lint]
extend-select = ["I", "TID252", "SIM"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
python = "^3.11"
discord-py = "^2.4.0"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.34"

